+++ ouds_core/lib/components/control/ouds_control_item.dart
@@ -10,7 +10,6 @@
 //
 
 /// @nodoc
-library;
 
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';