+++ ouds_core/lib/components/switch/ouds_switch.dart
@@ -9,7 +9,7 @@
 // Software description: Flutter library of reusable graphical components
 //
 
-/// OudsSwitch
+/// {@category Switch}
 library;
 
 import 'package:flutter/material.dart';
@@ -26,6 +26,8 @@ import 'package:ouds_theme_contract/theme/tokens/components/ouds_switch_tokens.d
 ///
 /// [OUDS Switch Design Guidelines](https://unified-design-system.orange.com/472794e18/p/18acc0-switch)
 ///
+/// **Reference design version : 1.5.0**
+///
 /// Switches allow the user to toggle between two states, typically "on" and "off". It is represented as a slider that changes its position or color to indicate
 /// the current state. Switches are used to enable or disable features, options, or settings in an intuitive and visual manner.
 ///
@@ -40,7 +42,7 @@ import 'package:ouds_theme_contract/theme/tokens/components/ouds_switch_tokens.d
 ///  [onChanged] Callback triggered when the user selects this switch.
 /// If `null`, switch is disabled and non-interactive.
 ///
-/// ## You can use [OudsSwitch] like this :
+/// ### You can use [OudsSwitch] component in your project, customizing parameters as needed :
 ///
 /// ```dart
 ///  OudsSwitch(